v0.0.6 jwt implementation

+++ src/main/java/com/leeknow/application/service/CommonService.java
@@ -7,11 +7,11 @@ public interface CommonService<T> {
 
     Page<T> findAll(Pageable pageable);
 
-    T findById(Long id);
+    T findById(Integer id);
 
     T save(T object);
 
     T update(T object);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
